refactor: fix concat_iterator bug, deduplicate get() methods, split vlog tests
- Fix next_sst_idx bug in concat_iterator (mid+1 → lo+1)
- Deduplicate get()/get_with_kind_inner() via lookup_memtable/lookup_sst_raw helpers
- Delete dead resolve_vlog_value() (zero callers)
- Split 1510-line vlog_integration_tests.rs into 6 focused submodules
- Extract force_flush() test helper, replacing 47 duplicated flush patterns
- Remove vestigial tests/.gitkeep
- Fix stale test comments and add missing assertions
- Remove dead set_vlog() calls from SST lookup path
