+++ kv-engine/src/lib.rs
@@ -17,6 +17,3 @@ pub mod wal;
 
 #[cfg(test)]
 mod tests;
-
-#[cfg(test)]
-mod vlog_integration_tests;
