+++ kv-engine/src/tests.rs
@@ -12,3 +12,4 @@ mod scan_flush;
 mod simple_leveled_compaction;
 mod sst;
 mod tiered_compaction;
+mod vlog_integration_tests;
